Nodes can be added in a user-friendly way.
The add script pretends to handle with a list of equal items instead of a tree and provides a user interface known from text processing applications: The user can e.g add a secondary heading at any place and the script will rearrange the other nodes to fulfill the user's expectations. See Node::addAsElementAfter() for details.